+++ src/js/core/scrollspy.js
@@ -135,9 +135,6 @@ export default {
             trigger(el, inview ? 'inview' : 'outview');
 
             state.inview = inview;
-
-            // change to `visibility: hidden` does not trigger observers
-            this.$update(el);
         },
     },
 };